Replaced long with long long for time stamp to calculate correct RSME

+++ src/ukf.h
@@ -66,7 +66,7 @@ public:
   bool is_initialized_;
 
   // previous timestamp
-  long previous_timestamp_;
+  long long previous_timestamp_;
 
   /**
    * Constructor
